fix multiple VMFactory component generation in multi-module projects
Previously, one ViewModelFactory component was generated in each module with a `ContributesViewModel`. This led to duplicate classes in projects with multiple modules and a confusing dependency graph.

The new approach only creates one single ViewModelFactory component in the same module (and package) where kotlin-inject-anvil's `MergeComponent` and `ContributesSubcomponent` are used. This makes a 1:1 mapping between components/subcomponents and the ViewModelFactory component associated with that scope.